build(pre-commit): update pre-commit to pass with no tests
