+++ .husky/pre-commit
@@ -2,5 +2,5 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 yarn lint
-yarn test -- --watch=false
+yarn test -- --watch=false --passWithNoTests
