+++ app/index.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import AppNavigator from "../features/navigation/AppNavigator";
-
+/* eslint-disable import/no-unused-modules */
 const AppEntry: React.FC = () => {
   return <AppNavigator />;
 };
 
 export default AppEntry;
+/* eslint-enable import/no-unused-modules */
