--- components/BasicButton.tsx
@@ -1,13 +1,0 @@
-import { Button } from "react-native"
-
-const BasicButton = ({ text, onPressHandle }: { 
-    text: string; 
-    onPressHandle: () => void 
-}) => (
-    <Button 
-        title={text} 
-        onPress={onPressHandle} 
-    />
-);
-  
-export default BasicButton;
