+++ features/navigation/AppNavigator.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import MatchCreationScreen from "../match-creation/presentation/MatchCreationScreen";
-import WarmupScreen from "../warmup/WarmupScreen";
+import WarmupScreen from "../warmup/presentation/WarmupScreen";
 import GameScreen from "../game/GameScreen";
 
 export type AppStackParamList = {
