--- features/timer/TimerScreenStyle.tsx
@@ -1,21 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      justifyContent: "center",
-      alignItems: "center",
-      backgroundColor: "#f5f5f5",
-      padding: 20,
-    },
-    timer: {
-      fontSize: 48,
-      fontWeight: "bold",
-      marginBottom: 30,
-    },
-    buttonContainer: {
-      flexDirection: "row",
-      justifyContent: "space-around",
-      width: "80%",
-    },
-  });
