+++ hooks/useThemeColor.ts
@@ -2,7 +2,7 @@
  * Learn more about light and dark modes:
  * https://docs.expo.dev/guides/color-schemes/
  */
-
+/* eslint-disable import/no-unused-modules */
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
@@ -19,3 +19,4 @@ export function useThemeColor(
     return Colors[theme][colorName];
   }
 }
+/* eslint-enable import/no-unused-modules */
