+++ package.json
@@ -57,6 +57,7 @@
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^8.57.0",
     "eslint-config-expo": "~8.0.1",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "gh": "^2.8.9",
     "husky": "^8.0.0",
     "jest": "^29.2.1",
